+++ scripts/helpers.sh
@@ -64,13 +64,6 @@ files_differ() {
 	! cmp -s "$1" "$2"
 }
 
-save_shell_history_option_on() {
-	local option_shell="$(get_tmux_option "$shell_history_option" "off")"
-	local option_bash="$(get_tmux_option "$bash_history_option" "off")"
-
-	[ "$option_shell" == "on" ] || [ "$option_bash" == "on" ]
-}
-
 get_grouped_sessions() {
 	local grouped_sessions_dump="$1"
 	export GROUPED_SESSIONS="${d}$(echo "$grouped_sessions_dump" | cut -f2 -d"$d" | tr "\\n" "$d")"
@@ -141,12 +134,6 @@ pane_contents_file_exists() {
 
 pane_contents_archive_file() {
 	echo "$(resurrect_dir)/pane_contents.tar.gz"
-}
-
-resurrect_history_file() {
-	local pane_id="$1"
-	local shell_name="$2"
-	echo "$(resurrect_dir)/${shell_name}_history-${pane_id}"
 }
 
 execute_hook() {