+++ scripts/restore.sh
@@ -303,25 +303,6 @@ restore_window_properties() {
 		done
 }
 
-restore_shell_history() {
-	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ { print $2, $3, $6, $9; }' $(last_resurrect_file) |
-		while IFS=$d read session_name window_number pane_index pane_command; do
-			if ! is_pane_registered_as_existing "$session_name" "$window_number" "$pane_index"; then
-				local pane_id="$session_name:$window_number.$pane_index"
-				local history_file="$(resurrect_history_file "$pane_id" "$pane_command")"
-
-				if [ "$pane_command" = "bash" ]; then
-					local read_command="history -r '$history_file'"
-					tmux send-keys -t "$pane_id" "$read_command" C-m
-				elif [ "$pane_command" = "zsh" ]; then
-					local accept_line="$(expr "$(zsh -i -c bindkey | grep -m1 '\saccept-line$')" : '^"\(.*\)".*')"
-					local read_command="fc -R '$history_file'; clear"
-					tmux send-keys -t "$pane_id" "$read_command" "$accept_line"
-				fi
-			fi
-		done
-}
-
 restore_all_pane_processes() {
 	if restore_pane_processes_enabled; then
 		local pane_full_command
@@ -389,10 +370,6 @@ main() {
 		restore_all_panes
 		handle_session_0
 		restore_window_properties >/dev/null 2>&1
-		execute_hook "pre-restore-history"
-		if save_shell_history_option_on; then
-			restore_shell_history
-		fi
 		execute_hook "pre-restore-pane-processes"
 		restore_all_pane_processes
 		# below functions restore exact cursor positions